Updates & Added source files
Source files were added from mimap library but had been modified to match
goals of pimap. Mimap is for working with priority but pimap is for working
with any value. Pimap does not have restrictions like values of items being
able to be sorted together or block having value/priority.

Pimap is subset of mimap, that means mimap will be modified to be based on it.
Changes dod not happen to source files but also README.md, it was modified
to also match changes made to source files.

Problem now is how pimap will be integrated into mimap.