Rename: 'pimap' -> 'pemap'
Project name 'pimap' does fail to upload on pypi. The error mentions that
projetc name is similar to existing project. It may be similar to 'mimap'
or 'pymap'.

Only rename and changing of occurences of 'pimap' were changed and nothing
else. The same will b edone again if the new project name also fails to be
uploaded.